nt::options
Struct PubSubOptions

pub struct PubSubOptions { /* private fields */ }

NetworkTables publish/subscribe options.

Implementations

impl PubSubOptions

pub fn poll_storage(self, poll_storage: u32) -> Self

Polling storage size for a subscription. Specifies the maximum number of\n updates NetworkTables should store between calls to the subscriberโs\n ReadQueue() function. If zero, defaults to 1 if sendAll is false, 20 if\n sendAll is true.

pub fn periodic(self, periodic: Duration) -> Self

How frequently changes will be sent over the network, in seconds.\n NetworkTables may send more frequently than this (e.g. use a combined\n minimum period for all values) or apply a restricted range to this value.\n The default is 100 ms.

pub fn exclude_publisher<T>(self, publisher: Publisher<'_, T>) -> Self

For subscriptions, if non-zero, value updates for ReadQueue() are not\n queued for this publisher.

pub fn send_all(self, send_all: bool) -> Self

Send all value changes over the network.

pub fn topics_only(self, topics_only: bool) -> Self

For subscriptions, donโt ask for value changes (only topic announcements).

pub fn prefix_match(self, prefix_match: bool) -> Self

Perform prefix match on subscriber topic names. Is ignored/overridden by\n Subscribe() functions; only present in struct for the purposes of getting\n information about subscriptions.

pub fn keep_duplicates(self, keep_duplicates: bool) -> Self

Preserve duplicate value changes (rather than ignoring them).

pub fn disable_remote(self, disable_remote: bool) -> Self

For subscriptions, if remote value updates should not be queued for\n ReadQueue(). See also disableLocal.

pub fn disable_local(self, disable_local: bool) -> Self

For subscriptions, if local value updates should not be queued for\n ReadQueue(). See also disableRemote.

pub fn exclude_self(self, exclude_self: bool) -> Self

For entries, donโt queue (for ReadQueue) value updates for the entryโs\n internal publisher.

pub fn hidden(self, hidden: bool) -> Self

For subscriptions, donโt share the existence of the subscription with the\n network. Note this means updates will not be received from the network\n unless another subscription overlaps with this one, and the subscription\n will not appear in metatopics.